Module Openai.Embeddings

type embeddings_input = {

model : string;

input : string list;

}

val jsonaf_of_embeddings_input : embeddings_input -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val embeddings_input_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> embeddings_input

val sexp_of_embeddings_input : embeddings_input -> Sexplib0.Sexp.t

val embeddings_input_of_sexp : Sexplib0.Sexp.t -> embeddings_input

val bin_shape_embeddings_input : Bin_prot.Shape.t

val bin_size_embeddings_input : embeddings_input Bin_prot.Size.sizer

val bin_write_embeddings_input : embeddings_input Bin_prot.Write.writer

val bin_writer_embeddings_input : embeddings_input Bin_prot.Type_class.writer

val bin_read_embeddings_input : embeddings_input Bin_prot.Read.reader

val __bin_read_embeddings_input__ : (int -> embeddings_input) Bin_prot.Read.reader

val bin_reader_embeddings_input : embeddings_input Bin_prot.Type_class.reader

val bin_embeddings_input : embeddings_input Bin_prot.Type_class.t

type response = {

data : embedding list;

}

Type definition for the response from the embeddings API.

and embedding = {

embedding : float list;

index : int;

}

Type definition for an individual embedding in the response.

val jsonaf_of_response : response -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val jsonaf_of_embedding : embedding -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val response_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> response

val embedding_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> embedding

val sexp_of_response : response -> Sexplib0.Sexp.t

val sexp_of_embedding : embedding -> Sexplib0.Sexp.t

val response_of_sexp : Sexplib0.Sexp.t -> response

val embedding_of_sexp : Sexplib0.Sexp.t -> embedding

val bin_shape_response : Bin_prot.Shape.t

val bin_shape_embedding : Bin_prot.Shape.t

val bin_size_response : response Bin_prot.Size.sizer

val bin_write_response : response Bin_prot.Write.writer

val bin_writer_response : response Bin_prot.Type_class.writer

val bin_size_embedding : embedding Bin_prot.Size.sizer

val bin_write_embedding : embedding Bin_prot.Write.writer

val bin_writer_embedding : embedding Bin_prot.Type_class.writer

val bin_read_response : response Bin_prot.Read.reader

val __bin_read_response__ : (int -> response) Bin_prot.Read.reader

val bin_reader_response : response Bin_prot.Type_class.reader

val bin_read_embedding : embedding Bin_prot.Read.reader

val __bin_read_embedding__ : (int -> embedding) Bin_prot.Read.reader

val bin_reader_embedding : embedding Bin_prot.Type_class.reader

val bin_response : response Bin_prot.Type_class.t

val bin_embedding : embedding Bin_prot.Type_class.t

val post_openai_embeddings : _ Eio.Net.t -> input:string list -> response

post_openai_embeddings ~input net makes an HTTP POST request to the OpenAI API embeddings endpoint with the given input and net.

It returns the parsed response as a response record.