Module Responses.Input_message

type role =

| User

| Assistant

| System

| Developer

val jsonaf_of_role : role -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val role_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> role

val sexp_of_role : role -> Sexplib0.Sexp.t

val role_of_sexp : Sexplib0.Sexp.t -> role

val bin_shape_role : Bin_prot.Shape.t

val bin_size_role : role Bin_prot.Size.sizer

val bin_write_role : role Bin_prot.Write.writer

val bin_writer_role : role Bin_prot.Type_class.writer

val bin_read_role : role Bin_prot.Read.reader

val __bin_read_role__ : (int -> role) Bin_prot.Read.reader

val bin_reader_role : role Bin_prot.Type_class.reader

val bin_role : role Bin_prot.Type_class.t

val role_to_string : role -> string

val role_of_string : string -> role

type text_input = {

text : string;

_type : string;

}

val jsonaf_of_text_input : text_input -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val text_input_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> text_input

val sexp_of_text_input : text_input -> Sexplib0.Sexp.t

val text_input_of_sexp : Sexplib0.Sexp.t -> text_input

val bin_shape_text_input : Bin_prot.Shape.t

val bin_size_text_input : text_input Bin_prot.Size.sizer

val bin_write_text_input : text_input Bin_prot.Write.writer

val bin_writer_text_input : text_input Bin_prot.Type_class.writer

val bin_read_text_input : text_input Bin_prot.Read.reader

val __bin_read_text_input__ : (int -> text_input) Bin_prot.Read.reader

val bin_reader_text_input : text_input Bin_prot.Type_class.reader

val bin_text_input : text_input Bin_prot.Type_class.t

type image_detail =

| High

| Low

| Auto

val jsonaf_of_image_detail : image_detail -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val image_detail_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> image_detail

val sexp_of_image_detail : image_detail -> Sexplib0.Sexp.t

val image_detail_of_sexp : Sexplib0.Sexp.t -> image_detail

val bin_shape_image_detail : Bin_prot.Shape.t

val bin_size_image_detail : image_detail Bin_prot.Size.sizer

val bin_write_image_detail : image_detail Bin_prot.Write.writer

val bin_writer_image_detail : image_detail Bin_prot.Type_class.writer

val bin_read_image_detail : image_detail Bin_prot.Read.reader

val __bin_read_image_detail__ : (int -> image_detail) Bin_prot.Read.reader

val bin_reader_image_detail : image_detail Bin_prot.Type_class.reader

val bin_image_detail : image_detail Bin_prot.Type_class.t

type image_input = {

image_url : string;

detail : string;

_type : string;

}

val jsonaf_of_image_input : image_input -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val image_input_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> image_input

val sexp_of_image_input : image_input -> Sexplib0.Sexp.t

val image_input_of_sexp : Sexplib0.Sexp.t -> image_input

val bin_shape_image_input : Bin_prot.Shape.t

val bin_size_image_input : image_input Bin_prot.Size.sizer

val bin_write_image_input : image_input Bin_prot.Write.writer

val bin_writer_image_input : image_input Bin_prot.Type_class.writer

val bin_read_image_input : image_input Bin_prot.Read.reader

val __bin_read_image_input__ : (int -> image_input) Bin_prot.Read.reader

val bin_reader_image_input : image_input Bin_prot.Type_class.reader

val bin_image_input : image_input Bin_prot.Type_class.t

type content_item =

| Text of text_input

| Image of image_input

val jsonaf_of_content_item : content_item -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val content_item_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> content_item

val sexp_of_content_item : content_item -> Sexplib0.Sexp.t

val content_item_of_sexp : Sexplib0.Sexp.t -> content_item

val bin_shape_content_item : Bin_prot.Shape.t

val bin_size_content_item : content_item Bin_prot.Size.sizer

val bin_write_content_item : content_item Bin_prot.Write.writer

val bin_writer_content_item : content_item Bin_prot.Type_class.writer

val bin_read_content_item : content_item Bin_prot.Read.reader

val __bin_read_content_item__ : (int -> content_item) Bin_prot.Read.reader

val bin_reader_content_item : content_item Bin_prot.Type_class.reader

val bin_content_item : content_item Bin_prot.Type_class.t

type content = content_item list

val jsonaf_of_content : content -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val content_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> content

val sexp_of_content : content -> Sexplib0.Sexp.t

val content_of_sexp : Sexplib0.Sexp.t -> content

val bin_shape_content : Bin_prot.Shape.t

val bin_size_content : content Bin_prot.Size.sizer

val bin_write_content : content Bin_prot.Write.writer

val bin_writer_content : content Bin_prot.Type_class.writer

val bin_read_content : content Bin_prot.Read.reader

val __bin_read_content__ : (int -> content) Bin_prot.Read.reader

val bin_reader_content : content Bin_prot.Type_class.reader

val bin_content : content Bin_prot.Type_class.t

type t = {

role : role;

content : content;

_type : string;

}

include Ppx_jsonaf_conv_lib.Jsonafable.S with type t := t

val t_of_jsonaf : Jsonaf_kernel__.Type.t -> t

val jsonaf_of_t : t -> Jsonaf_kernel__.Type.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t