Module Responses.Output_message

type role =

| Assistant

val jsonaf_of_role : role -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val role_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> role

val sexp_of_role : role -> Sexplib0.Sexp.t

val role_of_sexp : Sexplib0.Sexp.t -> role

val bin_shape_role : Bin_prot.Shape.t

val bin_size_role : role Bin_prot.Size.sizer

val bin_write_role : role Bin_prot.Write.writer

val bin_writer_role : role Bin_prot.Type_class.writer

val bin_read_role : role Bin_prot.Read.reader

val __bin_read_role__ : (int -> role) Bin_prot.Read.reader

val bin_reader_role : role Bin_prot.Type_class.reader

val bin_role : role Bin_prot.Type_class.t

type content = {

annotations : Annotation.t list;

text : string;

_type : string;

}

val jsonaf_of_content : content -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val content_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> content

val sexp_of_content : content -> Sexplib0.Sexp.t

val content_of_sexp : Sexplib0.Sexp.t -> content

val bin_shape_content : Bin_prot.Shape.t

val bin_size_content : content Bin_prot.Size.sizer

val bin_write_content : content Bin_prot.Write.writer

val bin_writer_content : content Bin_prot.Type_class.writer

val bin_read_content : content Bin_prot.Read.reader

val __bin_read_content__ : (int -> content) Bin_prot.Read.reader

val bin_reader_content : content Bin_prot.Type_class.reader

val bin_content : content Bin_prot.Type_class.t

type t = {

role : role;

id : string;

content : content list;

status : string;

_type : string;

}

include Ppx_jsonaf_conv_lib.Jsonafable.S with type t := t

val t_of_jsonaf : Jsonaf_kernel__.Type.t -> t

val jsonaf_of_t : t -> Jsonaf_kernel__.Type.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t