Module Responses.Response

type t = {

id : string;

object_ : string;

created_at : int;

status : Status.t;

error : Error.t option;

incomplete_details : Incomplete_details.t option;

instructions : string option;

max_output_tokens : int option;

model : string;

output : Item.t list;

parallel_tool_calls : bool option;

previous_response_id : string option;

reasoning : Request.Reasoning.t option;

store : bool option;

temperature : float option;

text : Text_cfg.t option;

tool_choice : Tool_choice.t option;

tools : Request.Tool.t list option;

top_p : float option;

truncation : string option;

usage : Usage.t option;

user : string option;

metadata : Metadata.t option;

}

include Ppx_jsonaf_conv_lib.Jsonafable.S with type t := t

val t_of_jsonaf : Jsonaf_kernel__.Type.t -> t

val jsonaf_of_t : t -> Jsonaf_kernel__.Type.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t