Module Responses.Response_stream

module Item : sig ... end

module Output_item_added : sig ... end

module Output_item_done : sig ... end

module Output_text_delta : sig ... end

module Output_text_done : sig ... end

module Reasoning_summary_text_delta : sig ... end

module Function_call_arguments_delta : sig ... end

module Function_call_arguments_done : sig ... end

module Response_created : sig ... end

module Response_in_progress : sig ... end

module Response_completed : sig ... end

module Response_incomplete : sig ... end

module Response_failed : sig ... end

module Part : sig ... end

module Content_part_added : sig ... end

module Content_part_done : sig ... end

module Response_refusal_delta : sig ... end

module Response_refusal_done : sig ... end

module File_search_call_in_progress : sig ... end

module File_search_call_searching : sig ... end

module File_search_call_completed : sig ... end

module Web_search_call_in_progress : sig ... end

module Web_search_call_searching : sig ... end

module Web_search_call_completed : sig ... end

module Unknown : sig ... end

type t =

| Output_item_added of Output_item_added.t

| Output_item_done of Output_item_done.t

| Output_text_delta of Output_text_delta.t

| Output_text_done of Output_text_done.t

| Function_call_arguments_delta of Function_call_arguments_delta.t

| Function_call_arguments_done of Function_call_arguments_done.t

| Response_created of Response_created.t

| Response_in_progress of Response_in_progress.t

| Reasoning_summary_text_delta of Reasoning_summary_text_delta.t

| Response_completed of Response_completed.t

| Response_incomplete of Response_incomplete.t

| Response_failed of Response_failed.t

| Content_part_added of Content_part_added.t

| Content_part_done of Content_part_done.t

| Response_refusal_delta of Response_refusal_delta.t

| Response_refusal_done of Response_refusal_done.t

| Annotation_added of Annotation_added.t

| File_search_call_in_progress of File_search_call_in_progress.t

| File_search_call_searching of File_search_call_searching.t

| File_search_call_completed of File_search_call_completed.t

| Web_search_call_in_progress of Web_search_call_in_progress.t

| Web_search_call_searching of Web_search_call_searching.t

| Web_search_call_completed of Web_search_call_completed.t

| Error of Error.t

| Unknown of Unknown.t

include Ppx_jsonaf_conv_lib.Jsonafable.S with type t := t

val t_of_jsonaf : Jsonaf_kernel__.Type.t -> t

val jsonaf_of_t : t -> Jsonaf_kernel__.Type.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t